Refactor: Contentful Rich Text display logic
