+++ src/utils/rich-text-options/embedded-rich-text-options.jsx
@@ -1,23 +1,33 @@
+/* eslint-disable no-underscore-dangle */
 /* eslint-disable react/destructuring-assignment */
-import React from 'react';
 import { documentToReactComponents } from '@contentful/rich-text-react-renderer';
 import { BLOCKS } from '@contentful/rich-text-types';
 import Image from 'next/image';
-import RICHTEXT_OPTIONS from '../rich-text-general-options';
-import Code from '../../components/code';
+import React from 'react';
 import getEntryMap from '../getEntryMap';
+import RICHTEXT_OPTIONS from './general-rich-text-options';
+import Code from '../../components/code';
 
-function renderOptions(links) {
+const EMBEDDED_RICHTEXT_OPTIONS = (links) => {
   const entryMap = getEntryMap(links.entries.block);
-  const assetMap = getEntryMap(links.assets.block);
-  const aboutRenderNode = {
+  const assetMap = links.assets && getEntryMap(links.assets.block);
+  return {
     [BLOCKS.EMBEDDED_ENTRY]: (node) => {
       const entry = entryMap.get(node.data.target.sys.id);
-      // eslint-disable-next-line no-underscore-dangle
       if (entry.__typename === 'CodeBlocks') {
         return (
           <div className="sm:col-span-10 space-y-3">
             <Code code={entry.codeBlockContent} language="javascript" />
+          </div>
+        );
+      }
+      if (entry.__typename === 'BasicDiv') {
+        return (
+          <div className="sm:col-span-10 space-y-3">
+            {documentToReactComponents(
+              entry.description.json,
+              RICHTEXT_OPTIONS
+            )}
           </div>
         );
       }
@@ -35,17 +45,5 @@ function renderOptions(links) {
       );
     },
   };
-
-  return {
-    renderNode: {
-      ...RICHTEXT_OPTIONS.renderNode,
-      ...aboutRenderNode,
-    },
-  };
-}
-
-function richTextOptions({ data: { json, links } }) {
-  return <>{documentToReactComponents(json, renderOptions(links))}</>;
-}
-
-export default richTextOptions;
+};
+export default EMBEDDED_RICHTEXT_OPTIONS;
